Refactoring to multi-mode maven project. closes #1
