--- pet-clinic-web/src/main/java/guru/springframework/App.java
@@ -1,13 +1,0 @@
-package guru.springframework;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
