+++ pet-clinic-web/src/main/resources/application.properties
